Introduce typing annotations
The goal is to initiate the effort of providing full annotations for the library.
Which is not yet the case, but we need to start small and progress
incrementally.
Hopefully this will create incentive from other contributors to add up
to this effort, even in tiny steps, until we can take advantage of
the multiple type checkers out there.